adds custom hooks for BLIs
