+++ frontend/src/components/UI/AllBLIRow/AllBLIRow.js
@@ -1,6 +1,5 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { useSelector } from "react-redux";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faClock } from "@fortawesome/free-regular-svg-icons";
 import CurrencyFormat from "react-currency-format";
@@ -15,6 +14,7 @@ import {
     totalBudgetLineAmountPlusFees,
 } from "../../../helpers/utils";
 import useGetUserFullNameFromId from "../../../helpers/useGetUserFullNameFromId";
+import { useIsBudgetLineEditableByStatus, useIsBudgetLineCreator } from "../../../helpers/useBudgetLines";
 
 /**
  * BLIRow component that represents a single row in the Budget Lines table.
@@ -35,15 +35,10 @@ const AllBLIRow = ({
 }) => {
     const [isExpanded, setIsExpanded] = React.useState(false);
     const [isRowActive, setIsRowActive] = React.useState(false);
-    const budgetLineCreator = useGetUserFullNameFromId(budgetLine?.created_by);
-    const loggedInUserId = useSelector((state) => state?.auth?.activeUser?.id);
-    const isBudgetLineDraft = budgetLine?.status === "DRAFT";
-    const isBudgetLineInReview = budgetLine?.status === "UNDER_REVIEW";
-    const isBudgetLinePlanned = budgetLine?.status === "PLANNED";
-    const isUserBudgetLineCreator = budgetLine?.created_by === loggedInUserId;
-    const isBudgetLineEditable =
-        (canUserEditBudgetLines || isUserBudgetLineCreator) &&
-        (isBudgetLineDraft || isBudgetLineInReview || isBudgetLinePlanned);
+    const budgetLineCreatorName = useGetUserFullNameFromId(budgetLine?.created_by);
+    const isUserBudgetLineCreator = useIsBudgetLineCreator(budgetLine);
+    const isBudgetLineEditableFromStatus = useIsBudgetLineEditableByStatus(budgetLine);
+    const isBudgetLineEditable = (canUserEditBudgetLines || isUserBudgetLineCreator) && isBudgetLineEditableFromStatus;
 
     // styles for the table row
     const removeBorderBottomIfExpanded = isExpanded ? "border-bottom-none" : "";
@@ -106,7 +101,7 @@ const AllBLIRow = ({
                     <dl className="font-12px">
                         <dt className="margin-0 text-base-dark">Created By</dt>
                         <dd id={`created-by-name-${budgetLine?.id}`} className="margin-0">
-                            {budgetLineCreator}
+                            {budgetLineCreatorName}
                         </dd>
                         <dt className="margin-0 text-base-dark display-flex flex-align-center margin-top-2">
                             <FontAwesomeIcon icon={faClock} className="height-2 width-2 margin-right-1" />
